extracted functional classes into their own packages
