+++ src/s1/t1/n2/exercici1/Types.java
@@ -1,0 +1,9 @@
+package s1.t1.n2.exercici1;
+
+interface Types {
+    void fotografiar();
+}
+
+interface Rellotge {
+    void alarma();
+}
