+++ src/s1/t1/n3/exercici1/Main.java
@@ -1,0 +1,8 @@
+package s1.t1.n3.exercici1;
+
+public class Main {
+    
+}
+
+
+
